Create PrimaryKeyField to serialize only the PKs of a relationship
Serialize the primary keys of a Model relationships is a standard
feature of REST libraries (see PrimaryKeyRelated type on Django REST).

This commit implements the PrimaryKeyField class. When defined for a
relationship attribute on a ModelSerializer, it will dump only a list of
IDs of the related models. The load occurs in the sameway, by given a
list of IDs.