Remove deprecated resource factories
resource_factory was a solution implemented to overcome the limitation
of flasgger (lib for swagger spec generation). Flasgger required one
Resource class per API entry. Since we had a single, generic class to
be used for any API entries, we created a factory to produce a different
class instance, nonetheless it has the same implementation.

The solution makes adds too much complexity to the code, and we intend
to generate swagger spec in another manner. As a result, we are
deprecating resource_factory.